Add EndpointTestCase
This class will help out testing endpoints by providing common methods.
